Break render details (mustache, fs) out into own file
