+++ lib/reportOnUserEvents.js
@@ -1,8 +1,7 @@
-const fs = require('fs');
-
-const Mustache = require('mustache');
-
 const describeTime = require('./describeTime');
+
+const renderUserEvent =  require('./renderUserEvents').renderUserEvent;
+const renderIndex = require('./renderUserEvents').renderIndex;
 
 const EventTypes = {
   CreateEvent: 'CreateEvent',
@@ -26,23 +25,12 @@ const hrefForCommitsFromPushEvent = (server, repo, pushEventPayload) => {
   return `${hrefForRepo(server, repo)}/compare/${before}...${head}`;
 };
 
-const render = (templateName, customData, event) => {
-  const baseData = {
-    actor: event.actor,
-    payload: event.payload
-  };
-  const renderData = Object.assign({}, baseData, customData);
-  const template = `${__dirname}/templates/userEvents/${templateName}.mustache`;
-  const content = fs.readFileSync(template);
-  return Mustache.render(content.toString(), renderData);
-};
-
 const buildUserEvent = (event, data, links) => {
   return {
     created_at: describeTime(event.created_at),
     links,
     repo: event.repo.name,
-    text: render(event.type, data, event)
+    text: renderUserEvent(event.type, data, event)
   }
 };
 
@@ -76,7 +64,6 @@ const EventTypeDescribers = {
     };
     return buildUserEvent(event, customData, links);
   }
-
 };
 
 const defaultDescription = (event, server) => {
@@ -93,15 +80,9 @@ const describeEvent = (event, server) => {
 };
 
 const reportOnUserEvents = (username, server, eventList) => {
-  const template = `${__dirname}/templates/userEvents/index.mustache`;
   const title = username;
   const userEvents = eventList.map((event) => describeEvent(event, server));
-  const data = {
-    userEvents: userEvents,
-    title
-  };
-  const content = fs.readFileSync(template);
-  return Mustache.render(content.toString(), data);
+  return renderIndex(title, userEvents);
 };
 
 module.exports = reportOnUserEvents;
